perf: Changed every variable to int and char[] back to string
Just learned that CPUs are optimized for ints and doubles. Program now will use a little bit more memory but be faster overall. Same with string.

+++ HumanCalculator/Interfaces.cs
@@ -8,16 +8,16 @@ namespace HumanCalculator
 {
     public interface ITime
     {
-        double[] Times { get; }
-        double AverageTime { get; }
+        int[] Times { get; }
+        int AverageTime { get; }
         void CalcAverageTime();
         string ShowAverageTime();
     }
     interface IScore
     {
         ITime Time { get; }
-        byte CurrentScore { get; }
-        void AddTime(double seconds);
+        int CurrentScore { get; }
+        void AddTime(int seconds);
         void AddScore();
     }
     interface IPlayer
